lint: clear golangci-lint findings across the carrier work
Fixes the issues golangci-lint (the cedar-style Lint job) flags, so the
golang-ccb PRs go green once they build against a tagged cedar:

- fstun/frame.go: drop unused frameOverhead const and (*frame).encodedLen.
- fstun/params.go: drop unused synParams.merge + minU16/32/64 (params are
  per-direction sender choices, not negotiated); fix the stale doc comment.
- fstun/conn.go: drop an unnecessary uint64() conversion (window already uint64).
- fstun/fsmode_test.go, bandwidth_test.go: use errors.Is(err, io.EOF) instead
  of == (errorlint).
- wscarrier/ws.go: close the WebSocket handshake response body (bodyclose).
- cmd/golang-ccb/main.go: extract the carrier config into buildCarrierSettings
  so run() is back under the gocyclo threshold.

No behavior change. `golangci-lint run ./...` is clean; transport suites and
the carrier e2e tests pass under -race.

+++ transport/fstun/conn.go
@@ -269,7 +269,7 @@ func (c *Conn) maybeAck() {
 	rd := c.recvData
 	c.recvMu.Unlock()
 	c.sendMu.Lock()
-	due := rd-c.lastAckSent >= uint64(c.params.window)/4
+	due := rd-c.lastAckSent >= c.params.window/4
 	c.sendMu.Unlock()
 	if due {
 		c.sendAck(rd)
